tests/cases/tree-construction/mensbeam03.dat

#data
<!DOCTYPE html><body></body> te st
#errors
(1,30): unexpected-char-after-body
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     " te st"

#data
X
#errors
#document-fragment
template
#document
| "X"

#data
<input type="text">
#errors
#document-fragment
form
#document
| <input>
|   type="text"

#data
<rb>X
#errors
#document-fragment
ruby
#document
| <rb>
|   "X"

#data
<!DOCTYPE html><ruby><div><rb>X</div></ruby>
#errors
(1,30): unexpected-parent-in-ruby
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <ruby>
|       <div>
|         <rb>
|           "X"

#data
<math definitionurl="http://example.com/"/>
#errors
#document-fragment
div
#document
| <math math>
|   definitionURL="http://example.com/"

#data
<!DOCTYPE html><body><form><template><div></form></div></template></form>
#errors
(1,49): unexpected-end-tag
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <form>
|       <template>
|         content
|           <div>

#data
<!DOCTYPE html><body><template><form><div></form></template>
#errors
(1,49): unexpected-end-tag
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <template>
|       content
|         <form>
|           <div>

#data
<dd><div></dd>
#errors
(1,14): unexpected-end-tag
#document-fragment
dl
#document
| <dd>
|   <div>

#data
<!DOCTYPE html><head></head></template>
#errors
(1,39): unexpected-end-tag
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>

#data
<!DOCTYPE html><select></template></select>
#errors
(1,34): unexpected-end-tag
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <select>

#data
<!DOCTYPE html><body><template><select><template>
#errors
(1,50): unexpected-eof
(1,50): unexpected-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <template>
|       content
|         <select>
|           <template>
|             content

#data
<tr>X
#errors
(1,5): fostered-char
#document-fragment
table
#document
| <tbody>
|   <tr>
| "X"

#data
<!DOCTYPE html><html xmlns="ook">
#errors
#document
| <!DOCTYPE html>
| <html>
|   xmlns="ook"
|   <head>
|   <body>
